third_party/blink/web_tests/platform/generic/fast/dom/Window/Location/window-shadow-location-using-string-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS result is 'pass.html'
